feature(user_product.proto): add fundamental service methods.
In this commit the fundamental methods for user, pocket and favorites manegement are added.
Also, there was a change in the submessages that used to belong to User, making the messages
more modular. verification_type and verification_status were added as attributes for User
message as it seems more natural to request them this way instead of requesting verification
information alone.